Небольшой рефакторинг кода
- Разбил длинный код на подмодули
- Подправил импорты
- Расставил пробелы и переносы строк
