+++ src/UI/primitives/values.py
@@ -1,0 +1,7 @@
+from PyQt5.QtGui import QColor, QFont
+
+DEFAULT_COLOR = QColor('red')
+DEFAULT_LINE_WIDTH = 3
+DEFAULT_POINT_SIZE = 21
+DEFAULT_FONT = QFont('Arial', 14, weight=QFont.Bold, italic=False)
+DEFAULT_PADDING = 20
