+++ src/controllers/Scenarios.py
@@ -1,27 +1,35 @@
 import logging
 import math
 import threading
-import time
-from math import pi, tan
-from math import sin
 from math import cos
+from math import pi
+from math import sin
 
-from PyQt5.QtCore import Qt, QEvent
+from PyQt5.QtCore import QEvent
 from PyQt5.QtGui import QColor
 
-from src.UI import UIPrimitives
+from src.UI.OverlayUI import OverlayUI, KeyboardKeys
+from src.UI.primitives import (
+    Circle,
+    Image,
+    Line,
+    Point,
+    Rect,
+    Text,
+)
+from src.UI.primitives.values import DEFAULT_FONT
+from src.controllers.Aspect import Aspect
+from src.controllers.ThaumInteractor import ThaumInteractor, createTI
 from src.logic.LinksGeneration import generateLinkMap
 from src.utils.LinkableValue import LinkableCoord, LinkableValue
-from src.UI.OverlayUI import OverlayUI, KeyboardKeys
-from src.controllers.ThaumInteractor import ThaumInteractor, createTI, Aspect
-from src.UI.UIPrimitives import Rect, Point, Line, Text, DEFAULT_FONT
 from src.utils.constants import MARGIN, THAUM_ASPECTS_INVENTORY_SLOTS_X, THAUM_ASPECTS_INVENTORY_SLOTS_Y, \
-    THAUM_HEXAGONS_SLOTS_COUNT, THAUM_ASPECT_RECIPES_CONFIG_PATH, THAUM_VERSION_CONFIG_PATH, \
-    THAUM_ADDONS_ASPECT_RECIPES_CONFIG_PATH
-from src.utils.utils import saveThaumControlsConfig, readJSONConfig, saveJSONConfig, eventsDelay, renderDelay, \
+    THAUM_HEXAGONS_SLOTS_COUNT, THAUM_ASPECT_RECIPES_CONFIG_PATH, THAUM_ADDONS_ASPECT_RECIPES_CONFIG_PATH
+from src.utils.utils import saveThaumControlsConfig, readJSONConfig, eventsDelay, renderDelay, \
     saveThaumVersionConfig, loadThaumVersionConfig
 
 pointTextAnchor = LinkableCoord(MARGIN, MARGIN)
+
+
 def enroll(UI: OverlayUI):
     UI.clearAll()
     UI.createExitButton()
@@ -46,7 +54,8 @@ def enroll(UI: OverlayUI):
     (cx, cy) = UI.getCenter()
     targetPoint = Point(cx, cy, color=QColor('yellow'))
     UI.addObject(targetPoint)
-    movablePoint = Point(MARGIN + MARGIN + DEFAULT_FONT.pointSize() * 2 * 16, MARGIN + 3.4 * (DEFAULT_FONT.pointSize() * 2), movable=True)
+    movablePoint = Point(MARGIN + MARGIN + DEFAULT_FONT.pointSize() * 2 * 16,
+                         MARGIN + 3.4 * (DEFAULT_FONT.pointSize() * 2), movable=True)
     UI.addObject(movablePoint)
     logging.info("Enrollment successfully showed")
 
@@ -89,13 +98,16 @@ def configureThaumWindowCoords(UI: OverlayUI):
     UI.addObject(Point(rectRB.x, rectRB.y, movable=True))
     logging.info("Configuring Thaum window rect dialogue successfully showed")
 
-    UI.setKeyCallback([KeyboardKeys.enter], confirmThaumWindowSlots,UI, rectThaumWindow.LT.x, rectThaumWindow.LT.y, rectThaumWindow.RB.x, rectThaumWindow.RB.y)
+    UI.setKeyCallback([KeyboardKeys.enter], confirmThaumWindowSlots, UI, rectThaumWindow.LT.x, rectThaumWindow.LT.y,
+                      rectThaumWindow.RB.x, rectThaumWindow.RB.y)
     UI.setKeyCallback([KeyboardKeys.backspace], enroll, UI)
+
 
 def confirmThaumWindowSlots(UI, LTx, LTy, RBx, RBy):
     thaumWindowWidth = RBx - LTx
     thaumWindowHeight = RBy - LTy
-    logging.info(f"Thaum window configured at: ({int(LTx)}, {int(LTy)}) x ({int(RBx)}, {int(RBy)}), width={thaumWindowWidth}, height={thaumWindowHeight}")
+    logging.info(
+        f"Thaum window configured at: ({int(LTx)}, {int(LTy)}) x ({int(RBx)}, {int(RBy)}), width={thaumWindowWidth}, height={thaumWindowHeight}")
 
     UI.clearAll()
     UI.createExitButton()
@@ -135,8 +147,10 @@ def confirmThaumWindowSlots(UI, LTx, LTy, RBx, RBy):
 
     rectAspectsLT = LinkableCoord(LTx + Ws * 0.25, LTy + Hs * 2.25)
     rectAspectsRB = LinkableCoord(LTx + Ws * 5.2, LTy + Hs * 7.25)
-    UI.addObject(Line(rectAspectsLT.x, rectAspectsLT.y, rectAspectsRB.x, rectAspectsRB.y, dashed=True, color=QColor('brown')))
-    UI.addObject(Line(rectAspectsRB.x, rectAspectsLT.y, rectAspectsLT.x, rectAspectsRB.y, dashed=True, color=QColor('brown')))
+    UI.addObject(
+        Line(rectAspectsLT.x, rectAspectsLT.y, rectAspectsRB.x, rectAspectsRB.y, dashed=True, color=QColor('brown')))
+    UI.addObject(
+        Line(rectAspectsRB.x, rectAspectsLT.y, rectAspectsLT.x, rectAspectsRB.y, dashed=True, color=QColor('brown')))
     rectAspectsListing = UI.addObject(
         Rect(rectAspectsLT.x, rectAspectsLT.y, rectAspectsRB.x, rectAspectsRB.y, dashed=True, color=QColor('lime')))
 
@@ -165,17 +179,23 @@ def confirmThaumWindowSlots(UI, LTx, LTy, RBx, RBy):
             horizontalListingLines[y - 1].S.y = yVal
             horizontalListingLines[y - 1].E.y = yVal
 
-    UI.addObject(Point(rectAspectsLT.x, rectAspectsLT.y, movable=True,onMoveCallback=updateListingRectCoords))
-    UI.addObject(Point(rectAspectsRB.x, rectAspectsRB.y, movable=True,onMoveCallback=updateListingRectCoords))  # aspects listing rectangle
+    UI.addObject(Point(rectAspectsLT.x, rectAspectsLT.y, movable=True, onMoveCallback=updateListingRectCoords))
+    UI.addObject(Point(rectAspectsRB.x, rectAspectsRB.y, movable=True,
+                       onMoveCallback=updateListingRectCoords))  # aspects listing rectangle
 
     aspectsScrollY = LinkableValue(LTy + Hs * 7.65)
-    pointAspectsScrollLeft = UI.addObject(Point(LTx + Ws * 2, aspectsScrollY, movable=True, color=QColor('lightblue')))  # aspects scroll left
-    pointAspectsScrollRight = UI.addObject(Point(LTx + Ws * 3.5, aspectsScrollY, movable=True, color=QColor('lightblue')))  # aspects scroll right
+    pointAspectsScrollLeft = UI.addObject(
+        Point(LTx + Ws * 2, aspectsScrollY, movable=True, color=QColor('lightblue')))  # aspects scroll left
+    pointAspectsScrollRight = UI.addObject(
+        Point(LTx + Ws * 3.5, aspectsScrollY, movable=True, color=QColor('lightblue')))  # aspects scroll right
 
     aspectsMixY = LinkableValue(LTy + Hs * 9)
-    pointAspectsMixLeft = UI.addObject(Point(LTx + Ws * 1, aspectsMixY, movable=True, color=QColor('pink')))  # aspects mix left
-    pointAspectsMixCreate = UI.addObject(Point(LTx + Ws * 2.75, aspectsMixY, movable=True, color=QColor('pink')))  # aspects mix create
-    pointAspectsMixRight = UI.addObject(Point(LTx + Ws * 4.5, aspectsMixY, movable=True, color=QColor('pink')))  # aspects mix right
+    pointAspectsMixLeft = UI.addObject(
+        Point(LTx + Ws * 1, aspectsMixY, movable=True, color=QColor('pink')))  # aspects mix left
+    pointAspectsMixCreate = UI.addObject(
+        Point(LTx + Ws * 2.75, aspectsMixY, movable=True, color=QColor('pink')))  # aspects mix create
+    pointAspectsMixRight = UI.addObject(
+        Point(LTx + Ws * 4.5, aspectsMixY, movable=True, color=QColor('pink')))  # aspects mix right
 
     rectLT = LinkableCoord(LTx + Ws * 2.55, LTy + Hs * 10.8)
     rectRB = LinkableCoord(LTx + Ws * 12.4, LTy + Hs * 14)
@@ -216,13 +236,16 @@ def confirmThaumWindowSlots(UI, LTx, LTy, RBx, RBy):
 
             deg30HexagonsLines[idx].S.x = rectHexagonsCC.x + cos(pi / 6) * rad + (i < 0) * i * slotSizeX
             deg30HexagonsLines[idx].E.x = rectHexagonsCC.x - cos(pi / 6) * rad + (i > 0) * i * slotSizeX
-            deg30HexagonsLines[idx].S.y = rectHexagonsCC.y + sin(pi / 6) * rad - (i < 0) * i * slotSizeY / 2 - (i > 0) * i * slotSizeY
-            deg30HexagonsLines[idx].E.y = rectHexagonsCC.y - sin(pi / 6) * rad - (i > 0) * i * slotSizeY / 2 - (i < 0) * i * slotSizeY
+            deg30HexagonsLines[idx].S.y = rectHexagonsCC.y + sin(pi / 6) * rad - (i < 0) * i * slotSizeY / 2 - (
+                    i > 0) * i * slotSizeY
+            deg30HexagonsLines[idx].E.y = rectHexagonsCC.y - sin(pi / 6) * rad - (i > 0) * i * slotSizeY / 2 - (
+                    i < 0) * i * slotSizeY
 
             deg60HexagonsLines[idx].S.x = -deg30HexagonsLines[idx].S.x + rectHexagonsCC.x * 2
             deg60HexagonsLines[idx].E.x = -deg30HexagonsLines[idx].E.x + rectHexagonsCC.x * 2
             deg60HexagonsLines[idx].S.y = deg30HexagonsLines[idx].S.y
             deg60HexagonsLines[idx].E.y = deg30HexagonsLines[idx].E.y
+
     updateHexagonsCoords()
 
     UI.addObject(Point(rectHexagonsCC.x, rectHexagonsCC.y, movable=True,
@@ -235,11 +258,13 @@ def confirmThaumWindowSlots(UI, LTx, LTy, RBx, RBy):
         saveThaumControlsConfig(pointWritingMaterials, pointPapers, rectAspectsListing.LT, rectAspectsListing.RB,
                                 pointAspectsScrollLeft, pointAspectsScrollRight,
                                 pointAspectsMixLeft, pointAspectsMixCreate, pointAspectsMixRight, rectInventory.LT,
-                                rectInventory.RB, rectHexagonsCC, (rectHexagonsCC.y - rectHexagonsTy) / (THAUM_HEXAGONS_SLOTS_COUNT // 2))
+                                rectInventory.RB, rectHexagonsCC,
+                                (rectHexagonsCC.y - rectHexagonsTy) / (THAUM_HEXAGONS_SLOTS_COUNT // 2))
         waitForCreatingTI(UI)
 
     UI.setKeyCallback([KeyboardKeys.enter], saveControls)
     UI.setKeyCallback([KeyboardKeys.backspace], configureThaumWindowCoords, UI)
+
 
 def chooseThaumVersion(UI: OverlayUI):
     UI.clearAll()
@@ -271,8 +296,8 @@ def chooseThaumVersion(UI: OverlayUI):
     for addonName in addonsRecipesConfig.keys():
         addonsSelectingState[addonName] = False
 
-    selectedVersionObject: list[Text|None] = [None]
-    selectedVersion: list[str|None] = [None]
+    selectedVersionObject: list[Text | None] = [None]
+    selectedVersion: list[str | None] = [None]
 
     (oldVersion, oldAddons) = loadThaumVersionConfig()
     logging.info(f"Selected in config version is: {oldVersion}. Selected in config addons is: {oldAddons}")
@@ -308,9 +333,11 @@ def chooseThaumVersion(UI: OverlayUI):
     curX = pointTextAnchor.x
     for i in range(len(versions)):
         version = versions[i]
+
         def onClickVersion(versionObject, version):
             logging.debug(f"Click on version {version}")
             selectVersion(versionObject, version)
+
         def selectVersion(versionObject, version):
             if selectedVersionObject[0] is not None:
                 selectedVersionObject[0].setColor(QColor('white'))
@@ -347,9 +374,11 @@ def chooseThaumVersion(UI: OverlayUI):
     startCurY = curY
     for i in range(len(addonsNames)):
         addonName = addonsNames[i]
+
         def onClickAddon(addonObject, addonName):
             logging.debug(f"Click on addon {addonName}")
             toggleAddonSelecting(addonObject, addonName)
+
         def toggleAddonSelecting(addonObject, addonName):
             addonsSelectingState[addonName] = not addonsSelectingState[addonName]
             if addonsSelectingState[addonName] is True:
@@ -358,6 +387,7 @@ def chooseThaumVersion(UI: OverlayUI):
             else:
                 addonObject.setColor(QColor('white'))
                 logging.debug(f"Deselected addon {addonName}")
+
         addonObject = UI.addObject(Text(
             0, 0,
             addonName,
@@ -398,12 +428,13 @@ def chooseThaumVersion(UI: OverlayUI):
     UI.setKeyCallback([KeyboardKeys.enter], onSumbit)
     UI.setKeyCallback([KeyboardKeys.backspace], configureThaumWindowCoords, UI)
 
+
 def beReadyForStartSolving(UI: OverlayUI):
     logging.info(f"Be ready for solving scenario started")
     UI.clearAll()
     UI.createExitButton()
 
-    UI.addObject(UIPrimitives.Text(
+    UI.addObject(Text(
         MARGIN, MARGIN,
         f"""Сейчас нейросеть будет определять аспекты, находящиеся на поле. 
 Необходимо подключение к интернету.
@@ -428,18 +459,18 @@ def waitForCreatingTI(UI: OverlayUI):
     UI.clearAll()
     UI.createExitButton()
 
-#     UI.addObject(Text(
-#         pointTextAnchor.x, pointTextAnchor.y,
-#         """Определяем аспекты в твоем столе.
-# Перенеси это окно так, чтобы оно не перекрывало окно с игрой
-# и нажми [Enter]
-# Чтобы вернуться назад, нажми [Backspace]""",
-#         color=QColor('white'),
-#         withBackground=True,
-#         backgroundColor=QColor('black'),
-#         padding=MARGIN,
-#         movable=True, UI=UI,
-#     ))
+    #     UI.addObject(Text(
+    #         pointTextAnchor.x, pointTextAnchor.y,
+    #         """Определяем аспекты в твоем столе.
+    # Перенеси это окно так, чтобы оно не перекрывало окно с игрой
+    # и нажми [Enter]
+    # Чтобы вернуться назад, нажми [Backspace]""",
+    #         color=QColor('white'),
+    #         withBackground=True,
+    #         backgroundColor=QColor('black'),
+    #         padding=MARGIN,
+    #         movable=True, UI=UI,
+    #     ))
 
     def startCreatingTI():
         UI.clearAll()
@@ -451,6 +482,7 @@ def waitForCreatingTI(UI: OverlayUI):
         # FIXME: Neurolink can't detect aspects on dark background
         # TI.updateAvailableAspectsInInventory()
         runResearching(UI, TI)
+
     # UI.setKeyCallback(KeyboardKeys.enter, startCreatingTI)
     # UI.setKeyCallback(KeyboardKeys.backspace, chooseThaumVersion, UI)
     startCreatingTI()
@@ -464,14 +496,15 @@ def runResearching(UI: OverlayUI, TI: ThaumInteractor):
     class Cell:
         x: int = None
         y: int = None
-        object: UIPrimitives.Circle = None
-        imageObject: UIPrimitives.Image = None
+        object: Circle = None
+        imageObject: Image = None
         aspect: Aspect = None
         isNone: bool = False
 
         def __init__(self, x: int, y: int):
             self.x = x
             self.y = y
+
         def __repr__(self):
             return f"Cell([{self.x}, {self.y}], aspect={self.aspect}, isNone={self.isNone})"
 
@@ -598,14 +631,14 @@ def runResearching(UI: OverlayUI, TI: ThaumInteractor):
         selectedCell[0] = None
         logging.debug(f'Exit cell state selecting dialogue. Showed only cells field')
 
-
     # draw clickable cells
     for ix in range(-THAUM_HEXAGONS_SLOTS_COUNT // 2 + 1, THAUM_HEXAGONS_SLOTS_COUNT // 2 + 1):
-        for iy in range(-THAUM_HEXAGONS_SLOTS_COUNT // 2 + (abs(ix) + 1) // 2 + 1, THAUM_HEXAGONS_SLOTS_COUNT // 2 - (abs(ix)) // 2 + 1):
+        for iy in range(-THAUM_HEXAGONS_SLOTS_COUNT // 2 + (abs(ix) + 1) // 2 + 1,
+                        THAUM_HEXAGONS_SLOTS_COUNT // 2 - (abs(ix)) // 2 + 1):
             hexagonCenterX = TI.rectHexagonsCC.x + ix * TI.hexagonSlotSizeX
             hexagonCenterY = TI.rectHexagonsCC.y + iy * TI.hexagonSlotSizeY - (ix % 2) * TI.hexagonSlotSizeY / 2
             cell = Cell(ix, iy)
-            cellObject = UIPrimitives.Circle(
+            cellObject = Circle(
                 hexagonCenterX,
                 hexagonCenterY,
                 r=TI.hexagonSlotSizeY / 2,
@@ -616,7 +649,7 @@ def runResearching(UI: OverlayUI, TI: ThaumInteractor):
             )
             cell.object = cellObject
             imageSide = TI.hexagonSlotSizeY / math.sqrt(2)
-            cellAspectImageObject = UIPrimitives.Image(
+            cellAspectImageObject = Image(
                 hexagonCenterX,
                 hexagonCenterY - imageSide,
                 imageSide,
@@ -632,7 +665,7 @@ def runResearching(UI: OverlayUI, TI: ThaumInteractor):
 
     # draw cell dialogue
     textYCoord = MARGIN
-    textCellIsNone = UI.addObject(UIPrimitives.Text(
+    textCellIsNone = UI.addObject(Text(
         MARGIN, textYCoord,
         'Ячейка недоступна (N)',
         color=QColor('white'),
@@ -646,7 +679,7 @@ def runResearching(UI: OverlayUI, TI: ThaumInteractor):
     ))
     cellSettingsStateDialogueObjects.append(textCellIsNone)
     textYCoord += textCellIsNone.h + MARGIN
-    textCellIsFree = UI.addObject(UIPrimitives.Text(
+    textCellIsFree = UI.addObject(Text(
         MARGIN, textYCoord,
         'Ячейка свободна (F)',
         color=QColor('white'),
@@ -664,7 +697,7 @@ def runResearching(UI: OverlayUI, TI: ThaumInteractor):
     textXCoord = MARGIN
     for i in range(len(TI.allAspects)):
         aspect = TI.allAspects[i]
-        textAspect = UI.addObject(UIPrimitives.Text(
+        textAspect = UI.addObject(Text(
             textXCoord, textYCoord,
             aspect.name,
             color=QColor('white'),
@@ -677,11 +710,11 @@ def runResearching(UI: OverlayUI, TI: ThaumInteractor):
             onClickCallbackArgs=[aspect],
             hoverable=True,
         ))
-        aspectImage = UI.addObject(UIPrimitives.Image(
+        aspectImage = UI.addObject(Image(
             textXCoord + MARGIN * 2, textYCoord,
             MARGIN * 2, MARGIN * 2,
             None,
-            ))
+        ))
         aspectImage.setImage(aspect.pixMapImage)
         cellSettingsStateDialogueObjects.append(textAspect)
         cellSettingsStateDialogueObjects.append(aspectImage)
@@ -692,7 +725,7 @@ def runResearching(UI: OverlayUI, TI: ThaumInteractor):
     UI.setObjectsVisibility(cellSettingsStateDialogueObjects, False)
 
     # draw base dialogue
-    activeStateTextObject = UI.addObject(UIPrimitives.Text(
+    activeStateTextObject = UI.addObject(Text(
         pointTextAnchor.x, pointTextAnchor.y,
         f"""Нейросеть определила аспекты на поле.
 Чтобы перегенерировать полученную цепочку решения, нажми [R]
@@ -714,7 +747,7 @@ def runResearching(UI: OverlayUI, TI: ThaumInteractor):
 
     def startPuttingLinkMap():
         UI.setAllObjectsVisibility(False)
-        onProcessText = UI.addObject(UIPrimitives.Text(
+        onProcessText = UI.addObject(Text(
             MARGIN, MARGIN,
             f"""Подожди, решение выкладывается на поле... 
 Не двигай мышью и не нажимай никакие кнопки.
@@ -743,9 +776,9 @@ def runResearching(UI: OverlayUI, TI: ThaumInteractor):
             goToNextSlot()
             UI.removeObject(onProcessText)
 
-        puttingAspectsThread = threading.Thread(target=startPuttingAspects)  # run in thread to not blocking keys callbacks
+        puttingAspectsThread = threading.Thread(
+            target=startPuttingAspects)  # run in thread to not blocking keys callbacks
         puttingAspectsThread.start()
-
 
     def goToNextSlot():
         logging.info("Going to next slot")
@@ -764,7 +797,7 @@ def runResearching(UI: OverlayUI, TI: ThaumInteractor):
         if isUnstoppableModeOn[0]:
             startPuttingLinkMap()
 
-    onPausedText = UI.addObject(UIPrimitives.Text(
+    onPausedText = UI.addObject(Text(
         MARGIN, MARGIN,
         f"""Программа проистановлена.
 
@@ -780,8 +813,10 @@ def runResearching(UI: OverlayUI, TI: ThaumInteractor):
 
     def switchToActiveState():
         logging.info("Switching to active state")
+
         def setUnstoppableModeOn():
             isUnstoppableModeOn[0] = True
+
         UI.clearKeyCallbacks()
         UI.setKeyCallback([KeyboardKeys.ctrl, KeyboardKeys.enter], setUnstoppableModeOn)
         UI.setKeyCallback([KeyboardKeys.enter], startPuttingLinkMap)
@@ -815,4 +850,3 @@ def runResearching(UI: OverlayUI, TI: ThaumInteractor):
     updateSolving()
     switchToActiveState()
     logging.debug("Hexagon field with configuring initial aspects showed")
-
